[Dy2St] Use OrderedSet in control flow name union
